+++ test/filters/test_colorize_syntax.rb
@@ -83,12 +83,8 @@ class Nanoc::Filters::ColorizeSyntaxTest < MiniTest::Unit::TestCase
     end
   end
 
-  def test_albino
-    if_have 'albino', 'nokogiri' do
-      if `which pygmentize`.strip.empty?
-        skip "could not find pygmentize"
-      end
-
+  def test_pygmentsrb
+    if_have 'pygments', 'nokogiri' do
       # Create filter
       filter = ::Nanoc::Filters::ColorizeSyntax.new
 
@@ -97,7 +93,7 @@ class Nanoc::Filters::ColorizeSyntaxTest < MiniTest::Unit::TestCase
       expected_output = '<pre title="moo"><code class="language-ruby"><span class="c1"># comment</span></code></pre>'
 
       # Run filter
-      actual_output = filter.run(input, :colorizers => { :ruby => :albino })
+      actual_output = filter.run(input, :colorizers => { :ruby => :pygmentsrb })
       assert_equal(expected_output, actual_output)
     end
   end